[In progress] Modifications for better files organization
