+++ cmd/image.go
@@ -1,7 +1,3 @@
-/*
-Copyright © 2023 Nicolas Sales <nicolas.cavalcante.dev@gmail.com>
-
-*/
 package cmd
 
 import (
@@ -11,13 +7,9 @@ import (
 // imageCmd represents the image command
 var imageCmd = &cobra.Command{
 	Use:   "image",
-	Short: "Image Engine for configuration filesystem",
-	Long: `A longer description that spans multiple lines and likely contains examples
-and usage of using your command. For example:
-
-Cobra is a CLI library for Go that empowers applications.
-This application is a tool to generate the needed files
-to quickly create a Cobra application.`,
+	Short: "image command for configuration filesystem and images",
+	Long: `The image engine command aims to sets up filesystem and images
+for the project.`,
 	Run: func(cmd *cobra.Command, args []string) {
 		cmd.Help()
 	},
