+++ cmd/init.go
@@ -1,12 +1,8 @@
-/*
-Copyright © 2023 NAME HERE <EMAIL ADDRESS>
-
-*/
 package cmd
 
 import (
 	"github.com/realnfcs/didactic-container/internal"
-	"github.com/realnfcs/didactic-container/internal/database"
+	"github.com/realnfcs/didactic-container/internal/models"
 	"github.com/spf13/cobra"
 )
 
@@ -18,7 +14,7 @@ var initCmd = &cobra.Command{
 	For that, is necessary to run the init command first to have the project ready to use.`,
 	Run: func(cmd *cobra.Command, args []string) {
 		internal.InitFolders()
-		database.CreateTable()
+        new(models.Image).CreateImageTable()
 	},
 }
